+++ packages/networks/src/endpoints.ts
@@ -2,30 +2,30 @@ import { UrlEndpoint } from './types'
 
 export const urlEndpointsMainnet: UrlEndpoint = {
   baseUrl: 'https://staking-api.injective.network:4444/api',
-  chainUrl: 'https://staking-evm.injective.network',
+  chainUrl: 'https://staking-grpc.injective.network',
   exchangeUrl: 'https://staking-api.injective.network:4444',
   explorerUrl: 'https://explorer-api.injective.network/v1',
 }
 
 export const urlEndpointsStaking: UrlEndpoint = {
   baseUrl: 'https://staking-api-testnet.injective.network:4444/api',
-  chainUrl: 'https://staking-evm-testnet.injective.network',
+  chainUrl: 'https://staking-grpc-testnet.injective.network',
   exchangeUrl: 'https://staking-api-testnet.injective.network:4444',
   explorerUrl: 'https://explorer-api-testnet.injective.network/v1',
 }
 
 export const urlEndpointsDevnet: UrlEndpoint = {
   baseUrl: 'https://devnet-api.injective.dev:4444/api',
-  chainUrl: 'https://devnet-evm.injective.dev',
+  chainUrl: 'https://devnet-grpc.injective.dev',
   exchangeUrl: 'https://devnet-api.injective.dev:4444',
   explorerUrl: 'https://devnet-explorer-api.injective.dev',
 }
 
-export const urlEndpointsInternal: UrlEndpoint = {
-  baseUrl: 'https://internal-api.injective.network/api',
-  chainUrl: 'https://internal-evm.injective.network',
-  exchangeUrl: 'https://internal-api.injective.network',
-  explorerUrl: 'https://internal-explorer-api.injective.dev',
+export const urlEndpointsPublic: UrlEndpoint = {
+  baseUrl: 'https://public-api.injective.network:4444/api',
+  chainUrl: 'https://public-grpc.injective.network',
+  exchangeUrl: 'https://public-api.injective.network:4444',
+  explorerUrl: 'https://public-explorer-api.injective.network',
 }
 
 export const urlEndpointsLocal: UrlEndpoint = {